Simplify and unify result tuples
This moves the predicate input value to the meta portion of the result
tuple. Previously it caused issues as it was hard to differentiate path
from input value.

In general, result tuples shouldn't use values as left-side of a tuple.
These positions should be reserved for defining types of information that
a tuple carry. ie `{:error, {:map, meta}}` uses `:map` to indicate it's
a map result.